+++ src/devcenter/azext_devcenter/aaz/latest/devcenter/admin/catalog/_wait.py
@@ -20,7 +20,7 @@ class Wait(AAZWaitCommand):
 
     _aaz_info = {
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/devcenters/{}/catalogs/{}", "2023-06-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/devcenters/{}/catalogs/{}", "2023-10-01-preview"],
         ]
     }
 
@@ -45,12 +45,22 @@ class Wait(AAZWaitCommand):
             help="The name of the catalog.",
             required=True,
             id_part="child_name_1",
+            fmt=AAZStrArgFormat(
+                pattern="^[a-zA-Z0-9][a-zA-Z0-9-_.]{2,62}$",
+                max_length=63,
+                min_length=3,
+            ),
         )
         _args_schema.dev_center_name = AAZStrArg(
             options=["-d", "--dev-center", "--dev-center-name"],
             help="The name of the dev center. Use `az configure -d dev-center=<dev_center_name>` to configure a default.",
             required=True,
             id_part="name",
+            fmt=AAZStrArgFormat(
+                pattern="^[a-zA-Z0-9][a-zA-Z0-9-]{2,25}$",
+                max_length=26,
+                min_length=3,
+            ),
         )
         _args_schema.resource_group = AAZResourceGroupNameArg(
             required=True,
@@ -126,7 +136,7 @@ class Wait(AAZWaitCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-06-01-preview",
+                    "api-version", "2023-10-01-preview",
                     required=True,
                 ),
             }
@@ -193,6 +203,9 @@ class Wait(AAZWaitCommand):
                 serialized_name="lastConnectionTime",
                 flags={"read_only": True},
             )
+            properties.last_sync_stats = AAZObjectType(
+                serialized_name="lastSyncStats",
+            )
             properties.last_sync_time = AAZStrType(
                 serialized_name="lastSyncTime",
                 flags={"read_only": True},
@@ -203,6 +216,31 @@ class Wait(AAZWaitCommand):
             )
             properties.sync_state = AAZStrType(
                 serialized_name="syncState",
+                flags={"read_only": True},
+            )
+            properties.sync_type = AAZStrType(
+                serialized_name="syncType",
+            )
+
+            last_sync_stats = cls._schema_on_200.properties.last_sync_stats
+            last_sync_stats.added = AAZIntType(
+                flags={"read_only": True},
+            )
+            last_sync_stats.removed = AAZIntType(
+                flags={"read_only": True},
+            )
+            last_sync_stats.synchronization_errors = AAZIntType(
+                serialized_name="synchronizationErrors",
+                flags={"read_only": True},
+            )
+            last_sync_stats.unchanged = AAZIntType(
+                flags={"read_only": True},
+            )
+            last_sync_stats.updated = AAZIntType(
+                flags={"read_only": True},
+            )
+            last_sync_stats.validation_errors = AAZIntType(
+                serialized_name="validationErrors",
                 flags={"read_only": True},
             )