+++ src/devcenter/azext_devcenter/aaz/latest/devcenter/admin/pool/_show.py
@@ -22,9 +22,9 @@ class Show(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2023-06-01-preview",
+        "version": "2023-10-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/projects/{}/pools/{}", "2023-06-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/projects/{}/pools/{}", "2023-10-01-preview"],
         ]
     }
 
@@ -49,12 +49,22 @@ class Show(AAZCommand):
             help="Name of the pool.",
             required=True,
             id_part="child_name_1",
+            fmt=AAZStrArgFormat(
+                pattern="^[a-zA-Z0-9][a-zA-Z0-9-_.]{2,62}$",
+                max_length=63,
+                min_length=3,
+            ),
         )
         _args_schema.project_name = AAZStrArg(
             options=["--project", "--project-name"],
             help="The name of the project. Use `az configure -d project=<project_name>` to configure a default.",
             required=True,
             id_part="name",
+            fmt=AAZStrArgFormat(
+                pattern="^[a-zA-Z0-9][a-zA-Z0-9-_.]{2,62}$",
+                max_length=63,
+                min_length=3,
+            ),
         )
         _args_schema.resource_group = AAZResourceGroupNameArg(
             required=True,
@@ -130,7 +140,7 @@ class Show(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-06-01-preview",
+                    "api-version", "2023-10-01-preview",
                     required=True,
                 ),
             }
@@ -193,6 +203,9 @@ class Show(AAZCommand):
                 serialized_name="devBoxDefinitionName",
                 flags={"required": True},
             )
+            properties.display_name = AAZStrType(
+                serialized_name="displayName",
+            )
             properties.health_status = AAZStrType(
                 serialized_name="healthStatus",
             )
@@ -208,6 +221,9 @@ class Show(AAZCommand):
                 serialized_name="localAdministrator",
                 flags={"required": True},
             )
+            properties.managed_virtual_network_regions = AAZListType(
+                serialized_name="managedVirtualNetworkRegions",
+            )
             properties.network_connection_name = AAZStrType(
                 serialized_name="networkConnectionName",
                 flags={"required": True},
@@ -216,8 +232,14 @@ class Show(AAZCommand):
                 serialized_name="provisioningState",
                 flags={"read_only": True},
             )
+            properties.single_sign_on_status = AAZStrType(
+                serialized_name="singleSignOnStatus",
+            )
             properties.stop_on_disconnect = AAZObjectType(
                 serialized_name="stopOnDisconnect",
+            )
+            properties.virtual_network_type = AAZStrType(
+                serialized_name="virtualNetworkType",
             )
 
             health_status_details = cls._schema_on_200.properties.health_status_details
@@ -230,6 +252,9 @@ class Show(AAZCommand):
             _element.message = AAZStrType(
                 flags={"read_only": True},
             )
+
+            managed_virtual_network_regions = cls._schema_on_200.properties.managed_virtual_network_regions
+            managed_virtual_network_regions.Element = AAZStrType()
 
             stop_on_disconnect = cls._schema_on_200.properties.stop_on_disconnect
             stop_on_disconnect.grace_period_minutes = AAZIntType(