+++ src/spring/azext_spring/_build_service.py
@@ -7,7 +7,7 @@
 
 import json
 from azure.cli.core.util import sdk_no_wait
-from .vendored_sdks.appplatform.v2023_09_01_preview import models
+from .vendored_sdks.appplatform.v2023_11_01_preview import models
 from ._buildservices_factory import BuildService
 from azure.cli.core.commands import LongRunningOperation
 from azure.cli.core.commands.client_factory import get_subscription_id