+++ src/aks-preview/HISTORY.rst
@@ -9,10 +9,26 @@ If there is no rush to release a new version, please just add a description of t
 
 To release a new version, please select a new version number (usually plus 1 to last patch version, X.Y.Z -> Major.Minor.Patch, more details in `\doc <https://semver.org/>`_), and then add a new section named as the new version number in this file, the content should include the new modifications and everything from the *Pending* section. Finally, update the `VERSION` variable in `setup.py` with this new version number.
 
+Pending
++++++++
+
+0.5.170
++++++++
+* Add `az aks approuting` and `az aks approuting zone` commands for managing App Routing.
+* Add `--node-provisioning-mode` to the `az aks update` command.
+* Add `--node-provisioning-mode` to the `az aks create` command.
+* Add Artifact Streaming enablement option to node pool property in `az aks nodepool add` and `az aks nodepool update`.
+* fix a bug in --support-plan handling when doing `az aks update`
+
 0.5.169
 +++++++
-* Vendor new SDK and bump API version to 2023-09-02-preview.
-* fix a bug in --support-plan handling when doing `az aks update`
+* Add `--network-plugin` to the `az aks update` command.
+* Add the KataCcIsolation option to --workload-runtime.
+* Update "VirtualMachines" agent pool type as Public Preview feature.
+* Add --disable-network-observability to `az aks update` cluster command.
+* Add `--node-soak-duration` to the `az aks nodepool add/update/upgrade` commands.
+* Add `--drain-timeout` to the `az aks nodepool add/update/upgrade` commands (already in [azure-cli](https://github.com/Azure/azure-cli/pull/27475)).
+
 
 0.5.168
 +++++++
@@ -20,6 +36,7 @@ To release a new version, please select a new version number (usually plus 1 to 
 
 0.5.167
 +++++++
+* Vendor new SDK and bump API version to 2023-09-02-preview.
 * Fix the default storagepool name value created for Azure Container Storage.
 * Ensure the correct nodepool name is picked and labelled by Azure Container Storage while installing with `az aks create`.